Fixed user & server creation on slackware (#93)
* Fixed user creation on Slackware

* Removed sudo from setPath function and updated line endings

* Cleaned up some mess

* I swear i already removed sudo from setPath

* Slight modifications to setPath function

* Updated description in setPath function

* Fixed syntax error

* Changed setPath messages to yellow and improved formatting

* Updated Minecraft crontabs
Added option to install minecraft crontabs

+++ easy-wi_install.sh
@@ -308,6 +308,29 @@ portRange() {
 
 }
 
+function setPath() {
+## As standard sudo users on Slackware do not have access to /sbin and /usr/sbin
+## directories which blocks access to usermod, userdel and useradd commands.
+
+CURRENTPATH=$(cat /etc/profile | grep PATH= | sed '/$PATH/d')
+CORRECTPATH='PATH="/usr/local/bin:/usr/bin:/bin:/usr/games:/sbin:/usr/sbin"'
+
+## If the PATH variable doesn't exist .. Yikes .. let's GTFO
+ yellowMessage " "
+ yellowMessage "Checking if PATH is set ...."
+ if [ $CURRENTPATH == "" ]; then
+ 	errorAndExit "No PATH detected, you need to fix this! ... Exiting"
+ 	exit 1
+ fi
+
+## Write the new PATH to /etc/profile if it's not correct already
+if [ $CURRENTPATH != $CORRECTPATH ]; then
+yellowMessage " "
+yellowMessage "Writing new PATH to /etc/profile ..."
+sed -i "s|$CURRENTPATH|"'PATH="/usr/local/bin:/usr/bin:/bin:/usr/games:/sbin:/usr/sbin"|' /etc/profile
+fi
+}
+
 cyanMessage " "
 yellowMessage "Please wait... Update is currently running."
 cyanMessage " "
@@ -344,10 +367,6 @@ elif [ -f /etc/centos-release ]; then
 elif [ -f /etc/os-release ]; then
 	INSTALLER="slackpkg"
 	OS="slackware"
-	#setenforce 0 >/dev/null 2>&1
-	#systemctl stop postfix
-	#systemctl disable postfix
-	#$INSTALLER clean all
 	$INSTALLER update
 	if [ -z "$(slackpkg search wget)" ]; then
 		checkInstall wget
@@ -1258,7 +1277,7 @@ if [ "$PHPINSTALL" == "Yes" ]; then
 		USE_PHP_VERSION='7.0'
 	elif [ "$OS" == "ubuntu" ] && [ "$OSVERSION" -ge "1610" ] && [ "$OSVERSION" -lt "1803" ]; then
 		USE_PHP_VERSION='7.1'
-	elif [ "$OS" == "ubuntu" ] && [ "$OSVERSION" -ge "1803" ]; then
+	elif [ "$OS" == "ubuntu" ] && [ "$OSVERSION" -ge "1803" ] && [ "$OSVERSION" -lt "2004" ]; then
 		USE_PHP_VERSION='7.2'
 	elif [ "$OS" == "ubuntu" ] && [ "$OSVERSION" -eq "2004" ]; then
 		USE_PHP_VERSION='7.4'
@@ -1852,6 +1871,13 @@ fi
 # No direct root access for masteruser. Only limited access through sudo
 if [ "$INSTALL" == "GS" ] || [ "$INSTALL" == "WR" ]; then
 	checkInstall sudo
+
+	if [ $OS == "slackware" ]; then
+		backUpFile /etc/profile
+		## Call the setPath function to check if sudo users have access to required programs
+		setPath
+	fi
+	
 	if [ -f /etc/sudoers ] && [ -z "$(grep "$MASTERUSER" /etc/sudoers | grep "$PKILL")" ]; then
 		echo "$MASTERUSER ALL = NOPASSWD: $PKILL" >>/etc/sudoers
 	fi
@@ -2018,7 +2044,7 @@ EOF
 
 			$INSTALLER -y install zlib1g
 			$INSTALLER -y install libc6-i386
-			if [ "$OS" == "debian" ] && [ "$OSVERSION" -gt "90" ] || [ "$OS" == "ubuntu" ] && [ "$OSVERSION" -gt "1803" ]; then
+			if [ "$OS" == "debian" ] && [ "$OSVERSION" -gt "9" ] || [ "$OS" == "ubuntu" ] && [ "$OSVERSION" -gt "1803" ]; then
 				$INSTALLER -y install lib32z1
 				$INSTALLER -y install lib32readline7
 				$INSTALLER -y install libreadline7:i386
@@ -2040,7 +2066,7 @@ EOF
 			$INSTALLER -y install libncursesw5-dev
 			$INSTALLER -y install zlib1g:i386
 		else
-			if [ "$OS" == "debian" ] && [ "$OSVERSION" -gt "90" ] || [ "$OS" == "ubuntu" ] && [ "$OSVERSION" -gt "1803" ]; then
+			if [ "$OS" == "debian" ] && [ "$OSVERSION" -gt "9" ] || [ "$OS" == "ubuntu" ] && [ "$OSVERSION" -gt "1803" ]; then
 				$INSTALLER -y install libreadline7 libncursesw5
 			else
 				$INSTALLER -y install libreadline5 libncursesw5
@@ -2106,52 +2132,64 @@ EOF
 
 	chown -cR "$MASTERUSER":"$MASTERUSER" /home/"$MASTERUSER" >/dev/null 2>&1
 
-	if [ "$OS" != "slackware" ]; then
-		if [ -f /etc/crontab ] && [ -z "$(grep 'Minecraft can easily produce 1GB' /etc/crontab)" ]; then
-			cyanMessage " "
-			okAndSleep "Installing Minecraft Crontabs"
-			if ionice -c3 true 2>/dev/null; then
-				IONICE="ionice -n 7 "
+	cyanMessage " "
+	cyanMessage "Minecraft cronjobs are used to periodically remove large log files"
+	cyanMessage "Do you want to install the minecraft cronjobs?"
+
+	OPTIONS=("Yes" "No" "Quit")
+	select OPTION in "${OPTIONS[@]}"; do
+		case "$REPLY" in
+		1 | 2) break ;;
+		3) errorAndQuit ;;
+		*) errorAndContinue ;;
+		esac
+	done
+
+	if [ $OPTION == "Yes" ]; then
+		if [ "$OS" != "slackware" ]; then
+			if [ -f /etc/crontab ] && [ -z "$(grep 'Minecraft can easily produce 1GB' /etc/crontab)" ]; then
+				cyanMessage " "
+				okAndSleep "Installing Minecraft Crontabs"
+				if ionice -c3 true 2>/dev/null; then
+					IONICE="ionice -n 7 "
+				fi
+
+				echo "#Minecraft can easily produce 1GB+ logs within one hour" >>/etc/crontab
+				echo "*/5 * * * * root nice -n +19 ionice -n 7 find /home/*/server/*/ -maxdepth 2 -type f -name \"screenlog.0\" -size +100M -delete" >>/etc/crontab
+				echo "# Even if sudo /usr/sbin/deluser --remove-all-files is used some data remain from time to time" >>/etc/crontab
+				echo "*/5 * * * * root nice -n +19 $IONICE find /home/ -maxdepth 2 -type d -nouser -delete" >>/etc/crontab
+				echo "*/5 * * * * root nice -n +19 $IONICE find /home/*/fdl_data/ /home/*/temp/ /tmp/ /var/run/screen/ -nouser -print0 | xargs -0 rm -rf" >>/etc/crontab
+				echo "*/5 * * * * root nice -n +19 $IONICE find /var/run/screen/ -maxdepth 1 -type d -nouser -print0 | xargs -0 rm -rf" >>/etc/crontab
 			fi
 
-			echo "#Minecraft can easily produce 1GB+ logs within one hour" >>/etc/crontab
-			echo "*/5 * * * * root nice -n +19 ionice -n 7 find /home/*/server/*/ -maxdepth 2 -type f -name \"screenlog.0\" -size +100M -delete" >>/etc/crontab
-			echo "# Even sudo /usr/sbin/deluser --remove-all-files is used some data remain from time to time" >>/etc/crontab
-			echo "*/5 * * * * root nice -n +19 $IONICE find /home/ -maxdepth 2 -type d -nouser -delete" >>/etc/crontab
-			echo "*/5 * * * * root nice -n +19 $IONICE find /home/*/fdl_data/ /home/*/temp/ /tmp/ /var/run/screen/ -nouser -print0 | xargs -0 rm -rf" >>/etc/crontab
-			echo "*/5 * * * * root nice -n +19 $IONICE find /var/run/screen/ -maxdepth 1 -type d -nouser -print0 | xargs -0 rm -rf" >>/etc/crontab
-		fi
-		echo " and 4"
-	elif [ "$OS" == "slackware" ]; then
+		elif [ "$OS" == "slackware" ]; then
 
-		if [ ! -f /etc/crond./easy-wi ]; then
-			touch /etc/cron.d/easy-wi
-		fi
-
-		if [ -f /etc/cron.d/easy-wi ] && [ -z "$(grep 'Minecraft can easily produce 1GB' /etc/cron.d/easy-wi)" ]; then
-			cyanMessage " "
-			okAndSleep "Installing Minecraft Crontabs"
-			if ionice -c3 true 2>/dev/null; then
-				IONICE="ionice -n 7 "
+			if [ ! -f /etc/crond./easy-wi ]; then
+				touch /etc/cron.d/easy-wi
 			fi
 
-			echo "#Minecraft can easily produce 1GB+ logs within one hour" >>/etc/cron.d/easy-wi
-			echo "*/5 * * * * root nice -n +19 ionice -n 7 find /home/*/server/*/ -maxdepth 2 -type f -name \"screenlog.0\" -size +100M -delete" >>/etc/cron.d/easy-wi
-			echo "# Even sudo /usr/sbin/deluser --remove-all-files is used some data remain from time to time" >>/etc/cron.d/easy-wi
-			echo "*/5 * * * * root nice -n +19 $IONICE find /home/ -maxdepth 2 -type d -nouser -delete" >>/etc/cron.d/easy-wi
-			echo "*/5 * * * * root nice -n +19 $IONICE find /home/*/fdl_data/ /home/*/temp/ /tmp/ /var/run/screen/ -nouser -print0 | xargs -0 rm -rf" >>/etc/cron.d/easy-wi
-			echo "*/5 * * * * root nice -n +19 $IONICE find /var/run/screen/ -maxdepth 1 -type d -nouser -print0 | xargs -0 rm -rf" >>/etc/cron.d/easy-wi
+			if [ -f /etc/cron.d/easy-wi ] && [ -z "$(grep 'Minecraft can easily produce 1GB' /etc/cron.d/easy-wi)" ]; then
+				cyanMessage " "
+				okAndSleep "Installing Minecraft Crontabs"
+				if ionice -c3 true 2>/dev/null; then
+					IONICE="ionice -n 7 "
+				fi
 
+				## Slackware does not use any screen socket directory, so cronjobs for /var/run/screen have been removed
+				echo "#Minecraft can easily produce 1GB+ logs within one hour" >>/etc/cron.d/easy-wi
+				echo "*/5 * * * * nice -n +19 ionice -n 7 find /home/*/server/*/ -maxdepth 2 -type f -name \"screenlog.0\" -size +100M -delete" >>/etc/cron.d/easy-wi
+				echo "# Even if sudo /usr/sbin/deluser --remove-all-files is used some data remain from time to time" >>/etc/cron.d/easy-wi
+				echo "*/5 * * * * nice -n +19 $IONICE find /home/ -maxdepth 2 -type d -nouser -delete" >>/etc/cron.d/easy-wi
+			fi
 		fi
-
-	fi
+	fi	
 
 	if [ "$OS" == "debian" ] || [ "$OS" == "ubuntu" ]; then
 		service cron restart 1>/dev/null
 	elif [ "$OS" == "centos" ]; then
 		systemctl restart crond.service 1>/dev/null
 	elif [ "$OS" == "slackware" ]; then
-		/etc/rc.d/rc.crond 1>/dev/null
+		/etc/rc.d/rc.crond restart 1>/dev/null
 	fi
 
 fi
